refactor: split ai.ts into AI abstraction layer + question-generate.ts
- ai.ts: pure AI abstraction (OpenAI client, structured/chat/streaming
  completions, model selection, env-configurable constants)
- question-generate.ts: all question domain logic (Zod schemas, prompt
  builders, MCQ/FRQ generation, FRQ grading, S3 persistence)
- tutor.ts: uses shared AI layer instead of own OpenAI client
- All models env-configurable: ADVANCED_MODEL, BASIC_MODEL, TUTOR_MODEL,
  GRADING_MODEL — no hardcoded model strings
- Removed OPENAI_URL fallback; single OPENAI_BASE_URL env var
- Zero functionality removed; all exports preserved
