+++ src/lib/server/services/question-cache.ts
@@ -1,7 +1,7 @@
 import { Question } from '$lib/server/models/question';
 import { SeenQuestion } from '$lib/server/models/seen-question';
 import { connectDb } from '$lib/server/db';
-import { generateAPQuestion, type GenerateResult } from './ai';
+import { generateAPQuestion, type GenerateResult } from './question-generate';
 import { logger } from '$lib/server/logger';
 import { runCacheMissClusterFlow } from './cache-miss-coordinator';
 import {
